Fix missing table on wb db query output
In #277 this code was changed in order to support a count number for
certain types of queries. That change had a side effect of removing the
default table that MySQL/MariaDB prints for regular queries because that
relies on TTY detection from MySQL/MariaDB itself. To revert to
preivious behavior, only run the new code if we know it is a row
modifying query that will just print the count. Otherwise, use the
previous code path.

Fixes #290